chore: migrate entities

+++ lib/core/domain/entities/loan_type.dart
@@ -22,7 +22,7 @@ class LoanType {
   final List<Currency> currencies;
   final int minDuration;
   final int maxDuration;
-  final int interestRate;
+  final double interestRate;
   final String image;
   final String createdAt;
   final String updatedAt;
@@ -35,7 +35,7 @@ class LoanType {
     List<Currency>? currencies,
     int? minDuration,
     int? maxDuration,
-    int? interestRate,
+    double? interestRate,
     String? image,
     String? createdAt,
     String? updatedAt,
